Build a better router

+++ controllers/Posts/delete.php
@@ -2,16 +2,29 @@
 
 use Core\Database;
 
-$config = require base_path("config.php");
+
+$config = require base_path('config.php');
 
 $pdo = new Database($config['database']);
+$currentUserId = 33;
+if ($_POST['_method'] === 'DELETE') {
+    $q = "SELECT * FROM posts where id = :id";
 
-$query = "DELETE FROM posts WHERE id = :id";
+    $id = $_POST['id'];
 
-$statement = $pdo->connection->prepare($query);
+    $post = $pdo->query($q, ['id' => $id])->findOrFail();
 
-$statement->bindParam(':id', $_GET['id']);
+    authorize($post['user_id'] === $currentUserId);
 
-$statement->execute();
+    $query = "DELETE FROM posts WHERE id = :id";
 
+    $statement = $pdo->connection->prepare($query);
 
+    $statement->bindParam(':id', $post['id']);
+
+    $statement->execute();
+
+    header("location: /posts");
+
+    exit();
+}
